% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ACutils_export.R
\name{ACheatmap_hcl}
\alias{ACheatmap_hcl}
\title{Heatmap of a matrix with hcl palettes.}
\usage{
ACheatmap_hcl(
  Mat,
  palette.name = "RdGy",
  col.n_breaks = 64,
  use_x11_device = TRUE,
  remove_diag = FALSE,
  title = "title",
  x_label = "x_axis",
  y_label = "y_axis",
  horizontal = TRUE
)
}
\arguments{
\item{Mat}{the matrix to be plotted.}

\item{palette.name}{the name of the hcl palette to be used. See here for possible names https://developer.r-project.org/Blog/public/2019/04/01/hcl-based-color-palettes-in-grdevices/}

\item{col.n_breaks}{the refinement of the palette.}

\item{use_x11_device}{boolean, if x11() device has to be activated or not.}

\item{remove_diag}{boolean, if the diagonal has to be removed or not.}

\item{title}{the title of the plot.}

\item{x_label}{the name of the x-axis in the plot.}

\item{y_label}{the name of the y-axis in the plot.}

\item{horizontal}{boolean, if the heatmap bar has to be horizontal or not.}
}
\value{
this function does not return anything
}
\description{
It plots the heatmap of a given matrix using one of the hcl palettes. It is a quicker version of the more complete function \code{\link{ACheatmap}}
}
