% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ACutils_export.R
\name{ACplot_graph}
\alias{ACplot_graph}
\title{Plot undirected graph from matrix or vector}
\usage{
ACplot_graph(
  G,
  col.links = "#3B9AB2",
  col.zeros = "grey95",
  graph.size = NULL,
  byrow = TRUE,
  main = " ",
  x_label = " ",
  y_label = " ",
  use_x11_device = TRUE
)
}
\arguments{
\item{G}{it may be (pxp)-matrix, where p is the number or nodes or a vector containing the elements of the adiacency
matrix. It is a vector, it may contain or not the diagonal elements. Its length should be equal to
\code{choose(p,2)} if diagonal elements are not included or \code{choose(p,2) + p}
if they are not included. Values should be 1 or 0.}

\item{col.links}{the color of the links.}

\item{col.zeros}{the color of the absent links, the zeros in the adiacency matrix.}

\item{graph.size}{the number of nodes. Used only is \code{G} is a vector.}

\item{byrow}{boolean, used only is \code{G} is a vector. Set TRUE if the elements of G are taken by row from its
adiacency matrix. Set FALSE if they are taken by column.}

\item{main}{the title of the plot.}

\item{x_label}{the name of the x-axis in the plot.}

\item{y_label}{the name of the y-axis in the plot.}

\item{use_x11_device}{boolean, if x11() device has to be activated or not.}
}
\value{
this function does not return anything
}
\description{
\loadmathjax Given a (pxp)-matrix or a n_elem-dimensional vector, it plots the corresponding undirected graph.
}
