% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ACutils_export.R
\name{ACplot_network}
\alias{ACplot_network}
\title{Plot undirected network from matrix or vector}
\usage{
ACplot_network(
  G,
  labels.name = NULL,
  col.nodes = NULL,
  col.links = NULL,
  col.labels = "white",
  node.size = 12,
  label.size = 3.5,
  use_x11_device = T
)
}
\description{
Plot undirected network from matrix or vector
}
