% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun-baseline.R
\name{baseline}
\alias{baseline}
\alias{baseline.data.frame}
\alias{baseline.grouped_df}
\title{subtract baseline from signal}
\usage{
baseline(x, variables, reference, method = mean, by_group)

\method{baseline}{data.frame}(x, variables, reference, method = mean,
  by_group)

\method{baseline}{grouped_df}(x, variables, reference, method = mean,
  by_group)
}
\arguments{
\item{x}{a \code{data.frame} or \code{grouped data frame}}

\item{variables}{character vector of variables (columns) to modify}

\item{reference}{logical predicate that identifies reference observations, given as string or bare;
will be passed to \code{\link{subset}}}

\item{method}{method with which to compute the baseline, mean or median; given as symbol or string}

\item{by_group}{character vector of variables to group by (\code{\link{dplyr::group_by}})
in case the baseline is to be taken from means of sets of observations
rather than from all observations directly}
}
\value{
a modified \code{data.frame}
}
\description{
Calculate baseline signal and subtract it from all observations.
}
\section{Methods (by class)}{
\itemize{
\item \code{data.frame}: computes mean/median of desired variables with \code{vapply}
and subtracts them from their respective variables using \code{mapply}

\item \code{grouped_df}: constructed by \code{\link{acutils::metamethod}}
saves attributes and grouping variables of \code{x}, captures original call
and modifies it to be run by \code{by}, then restores attributes
}}

\section{Grouped data frames (\code{dplyr} package)}{

The method for class \code{grouped_df} is home brewed as I don't know
how to properly handle this class.
}

