% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ACutils_export.R
\name{boxplot_from_matrix}
\alias{boxplot_from_matrix}
\title{Plot boxplot from matrix}
\usage{
boxplot_from_matrix(
  data,
  use_x11_device = TRUE,
  use_ggplot = TRUE,
  main = " ",
  x_label = " ",
  y_label = " "
)
}
\arguments{
\item{data}{an (nxp)-matrix.}

\item{use_x11_device}{boolean, if x11() device has to be activated or not.}

\item{use_ggplot}{if boxplots should be done using ggplot or not.}

\item{main}{the title of the plot.}

\item{x_label}{the name of the x-axis in the plot.}

\item{y_label}{the name of the y-axis in the plot.}
}
\value{
this function does not return anything
}
\description{
Given a (nxp)-matrix, it plots the boxplots of the columns.
}
