% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ACutils_export.R
\name{boxplot_from_vectors}
\alias{boxplot_from_vectors}
\title{Plot boxplot from vectors}
\usage{
boxplot_from_vectors(
  ...,
  names,
  use_x11_device = TRUE,
  use_ggplot = TRUE,
  title = "Title",
  x_label = "x_axis",
  y_label = "y_axis"
)
}
\arguments{
\item{...}{an arbitrary number of vectors. The function does not checks if they are really vectors, so be careful.}

\item{names}{the names of the groups of the different boxplots. Not defaulted for safety, can be set to 1:length(...).}

\item{use_x11_device}{boolean, if x11() device has to be activated or not.}

\item{title}{the title of the plot.}

\item{x_label}{the name of the x-axis in the plot.}

\item{y_label}{the name of the y-axis in the plot.}
}
\value{
this function does not return anything
}
\description{
Takes an arbitrary number of vectors and plot the boxplots. When calling this function, any arguments after ... must be fully named.
}
\examples{
boxplot_from_vectors(v1,v2,v3, names = 1:3)
boxplot_from_vectors(v1,v2,v3, names = c("one","two", "three"), title = "Nice Boxplot")

}
