% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun-clean_column_names.R
\name{clean_column_names}
\alias{clean_column_names}
\title{clean up superfluous characters in column names}
\usage{
clean_column_names(x, gate = "dummy", object = "Main")
}
\arguments{
\item{x}{a \code{data frame}}

\item{gate}{gate name to be cleared, given as a regular expression;
defaults to "dummy", which does nothing, unless there is a gate called "dummy"}

\item{object}{name of object to be cleared; defaults to "Main"}
}
\value{
data frame with modified column names
}
\description{
Remove gate and object names from column names. Some reformatting to make them easier to read.
}
\details{
ScanR export files are tab delimited tables whose columns contain the name of the object,
the gate it belongs to, the parameter in question and the measure being exported, e.g. mean or error.
The names are cumbersome and when loaded with \code{check.names = TRUE}, they can be quite confusing.
Usually most of these can be removed without any percievable loss.

This is a simple stream of gsub calls that removes superfluous information from the column names.
It also improves the formatting. \code{\link{make.names}} is run first.
}
