% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun-clip_data.R
\name{clip_data}
\alias{clip_data}
\alias{clip_data.default}
\alias{clip_data.factor}
\alias{clip_data.numeric}
\alias{clip_data.data.frame}
\title{remove extreme values from a vector}
\usage{
clip_data(x, clip = 1L, preserve.order = TRUE)

\method{clip_data}{default}(x, clip = 1L, preserve.order = TRUE)

\method{clip_data}{factor}(x, clip = 1L, preserve.order = TRUE)

\method{clip_data}{numeric}(x, clip = 1L, preserve.order = TRUE)

\method{clip_data}{data.frame}(x, clip = 1L, column)
}
\arguments{
\item{x}{a vector or a data frame}

\item{clip}{\strong{either} number(s) or fraction(s) (rounding up)
of values to be removed, see \code{Details}}

\item{preserve.order}{logical flag whether result should be returned
sorted or in the original order (default)}
}
\value{
A clipped vector or data frame.
}
\description{
Remove the smallest and greatest values from a vector of from a column of a data frame.
}
\details{
The \code{clip} arguments determines the number of items/observations removed:
\itemize{
\item{if clip >= 1, this number of observations will be removed}
\item{if clip < 1, this fraction of observations will be removed}
\item{if clip < 1 and of length 1, half the fraction will be removed
      from the bottom and half from the top of the range}
\item{if clip >= 1 and of length 1, that number of items will be removed
      from both bottom and top of the range}
\item{if clip is of length 2, the first value determines clipping
      at the bottom of the range and the second - at the top}
}

After clipping a t test is run to determine whether the clipping significantly
changed the mean of the vector. In case of a data frame this is also done for
every other numeric variable. The result of the test is reported as a warning
if positive (p < 0.05) and as a message if negative.
}
\section{Methods (by class)}{
\itemize{
\item \code{default}: default method

\item \code{factor}: coerces \code{x} to characer, calls the character method, and restores the original levels

\item \code{numeric}: adds a location check

\item \code{data.frame}: clips the specified column and test all numeric columns for changes
}}

\author{
Aleksander Chlebowski, June 2018
}
