% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun-construct.names.R
\name{construct.names}
\alias{construct.names}
\title{translate RSz construct names}
\usage{
construct.names(data, dictionary)
}
\arguments{
\item{data}{a \code{data.frame}; must contain a column called \code{cells}}

\item{dictionary}{optional dictionary file or object, see \code{Details}}
}
\value{
a modified data frame
}
\description{
Convert construct numbers in the RSz series to their contents (inserts).
}
\details{
The function utilizes a dictionary table to modify the "cells" column
in \code{data} accordingly. By default this is taken from a file in the package's
\code{data} directory, which can be replaced or modified at will as long as it
is a tab delimited .txt file with two columns: "cells" and "insert".

Alternatively a custom file or a data frame object can be provided. The same restrictions apply.

Empty cell lines can be (and indeed are) placed in the dictionary.
Their "insert" field is the same as their "cells" filed.
}
