% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ACutils_export.R
\name{dnct}
\alias{dnct}
\title{Non Central Student t - Density}
\usage{
dnct(x, n0, mu0, gamma0)
}
\arguments{
\item{x}{the point where to evaluate the density.}

\item{n0}{the degree of fredoom.}

\item{mu0}{the location parameter.}

\item{gamma0}{the scale parameter.}
}
\value{
scalar representing the evaluation of the density at x
}
\description{
\loadmathjax This function evaluates the density of a Non Central Student t in a given point. Such a distribution is defined as follow:
\mjsdeqn{X~\sim~nct(n_{0},\mu_{0},\gamma_{0})} if
\mjtdeqn{$$\begin{eqnarray*}f(X|n_{0},\mu_{0},\gamma_{0}) = \frac{\Gamma(\frac{n_{0}+1}{2})}{\Gamma(\frac{n_{0}}{2})}\frac{1}{\sqrt{\pi\gamma_{0}n_{0}}}\left(1+\frac{1}{n_{0}}(\frac{x-\mu_{0}}{\gamma})^{2}\right)^{-\frac{n_{0}+1}{2}} {eqnarray*}$$}{\begin{eqnarray*}f(X|n_{0},\mu_{0},\gamma_{0}) = \frac{\Gamma(\frac{n_{0}+1}{2})}{\Gamma(\frac{n_{0}}{2})}\frac{1}{\sqrt{\pi\gamma_{0}n_{0}}}\left(1+\frac{1}{n_{0}}(\frac{x-\mu_{0}}{\gamma})^{2}\right)^{-\frac{n_{0}+1}{2}}  \end{eqnarray*}}{\begin{eqnarray*} f(X|n_{0},\mu_{0},\gamma_{0}) = \frac{\Gamma(\frac{n_{0}+1}{2})}{\Gamma(\frac{n_{0}}{2})}\frac{1}{\sqrt{\pi\gamma_{0}n_{0}}}\left(1+\frac{1}{n_{0}}(\frac{x-\mu_{0}}{\gamma})^{2}\right)^{-\frac{n_{0}+1}{2}}  \end{eqnarray*}}
where \mjseqn{n_{0}} are the degree of freedom, \mjseqn{\mu_{0}} is the location parameter and \mjseqn{\gamma_{0}} is the scale parameter. The usual t density can be recovered by placing \mjseqn{\mu_{0}=0} and \mjseqn{\gamma_{0} = 1}.
}
