% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun-errors.R
\name{errors}
\alias{errors}
\alias{sem}
\alias{conf}
\title{error functions}
\usage{
sem(x, na.rm = TRUE)

conf(x, na.rm = TRUE, confidence = 0.99)
}
\arguments{
\item{x}{a numeric vector}

\item{na.rm}{logical flag whether to remove NA values}

\item{confidence}{the desired confidence}
}
\value{
a single numeric
}
\description{
Functions to express the uncertainty of a sample.
}
\details{
These are helper functions for drawing error bars on plots.
They both calculate a measure of uncertainty and return a single number,
which is the width of the "uncertainty interval".
Add and subtract this unmber to the sample mean to get
the position of the error bars.
}
\section{Functions}{
\itemize{
\item \code{sem}: standard error of the mean, i.e.
the standard deviation of the sample divided by
the square root of the number of elements

\item \code{conf}: half of the width of the confidence interval of the mean for a numeric vector
}}

\section{Disclaimer}{

The confidence interval is calculated with a t distribution.
I don't understand how the confidence interval is calculated.
}

