% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun-fetch_files.R
\name{fetch_files}
\alias{fetch_files}
\title{fetch ScanR result/export files}
\usage{
fetch_files(where.from, where.to, object = "Main")
}
\arguments{
\item{where.from}{directory where all screen data is stored;
defaults to current working directory,
otherwise must be an existing directory}

\item{where.to}{directory where all files will be copied;
defaults to current working directory,
otherwise may be any directory;
a non-existing one will be created}

\item{object}{regular expression that defines which ParameterData files to copy;
defaults to Main as this is always present and usually is the only relevant one}
}
\description{
Copy all ScanR result files, chosen parameter data files, and the screen log
into a single location.
}
\details{
Given the location of a master directory, i.e. one where multiple scans are stored,
this function goes into all scan directories and locates all files in the
"Population Data" subdirectory (including hidden ones), as well as any "ParameterData" files that
contain an object name specified by a regular expression (\code{object}).
All files are copied to a specified directory. Previously present files will be overrwritten.

Names of scan directories, which are usually plate numebrs,
are added to all respective file names.

If the master directory contains a file with a name matching the regex "screenlog",
it will also be copied.

Sort the copied files as you wish.
}
