% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun-get.stack.R
\name{get.stack}
\alias{get.stack}
\title{\code{get} within a function}
\usage{
get.stack(x)
}
\arguments{
\item{x}{a character string}
}
\value{
Returns the value of a variable named \code{x} if one exists
       anywhere in the current call stack. Otherwise throws an error.
}
\description{
Get object from the call stack rather than the search path.
}
\details{
This version of get searches the whole call stack but not the search path.
Basically a rewrite of \code{dynGet} (actually, rewritten based on \code{dynGet}).
}
