% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun-ggthemes.R
\name{ggthemes}
\alias{ggthemes}
\alias{center_titles}
\alias{rotate_labels}
\title{modify ggplot theme}
\usage{
center_titles()

rotate_labels()
}
\description{
Convenience functions to slightly modify ggplot objects.
}
\section{Functions}{
\itemize{
\item \code{center_titles}: called from global environment;
justifies plot titles and subtitles to center

\item \code{rotate_labels}: added to ggplot object;
rotates x axis text to vertical and adjusts position
}}

