% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ACutils_export.R
\name{Graph_distance}
\alias{Graph_distance}
\title{Compute number of errors}
\usage{
Graph_distance(G1, G2, diag = F)
}
\arguments{
\item{G1}{binary matrix, only the upper part is used.}

\item{G2}{binary matrix, only the upper part is used.}

\item{diag}{boolean, if the diagonal has to be checked or not}
}
\value{
the number of elements that are different
}
\description{
Takes two matrices containing 0s and 1s and compute the number of elements that are different. The diagonal may or may not be checked.
}
