% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun-interleave.R
\name{interleave}
\alias{interleave}
\title{interleave two vectors}
\usage{
interleave(x, y)
}
\arguments{
\item{x, y}{vectors}
}
\description{
Given two vectors, x and y, outputs a vector x1, y1, x2, y2, etc.
Currently only supports two vectors.
}
