% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ACutils_export.R
\name{KL_dist}
\alias{KL_dist}
\title{Kullback-Leibler for zero-mean Gaussian data}
\usage{
KL_dist(Ktr, K)
}
\arguments{
\item{Ktr}{the true matrix. Can be both covariance or precision}

\item{K}{the estimated matrix. Can be both covariance or precision}
}
\value{
scalar, the distance between N(0, Ktr) and N(0, K)
}
\description{
Compute the Kullback-Leibler for zero-mean Gaussian distributed data. Can be used to measure the distance between the true underlying covariance (or precision) matrix and the estimated one.
}
