% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun-metamethod-grouped_df.R
\name{metamethod}
\alias{metamethod}
\title{create S3 method for class \code{grouped_df}}
\usage{
metamethod(fun)
}
\arguments{
\item{fun}{a function; this must be an S3 method for class \code{data.frame};
also, its first argument must be named \code{x}}
}
\value{
a function: a method for class \code{grouped_df} (from package \code{dplyr})
}
\description{
Given an S3 method for class \code{data.frame}, create an S3 method
for class \code{grouped_df} for the same generic function.
}
\details{
This is a utility function for writing packages.
Functions that modify data frames by replacing or adding columns may fail
on grouped data frames so the class will require a dedicated method.
\code{metamethod} is a generalized constructor of such methods.

In all cases the original call to \code{foo} is captured with \code{match.call}
and reconstructed into a call to \code{by}. The first argument of \code{foo},
which must be named \code{x}, is passed to the \code{data} argument of the new call.
Grouping variables are isolated as a list and passed to the \code{INDICES} argument.
The \code{FUN} argument is set to \code{foo.data.frame},
i.e. the \code{x} of \bold{this} function.
The remaining arguments are passed as is.
}
\section{How to use}{

When editing an .R file for generic \code{foo} and its S3 methods,
simply add another entry where \code{foo.grouped_df} is defined as the result of
\code{metamethod} acting on \code{foo.data.frame}.

Copy the following \code{roxygen} tag before the function definition
to have it show up nicely in the documentation file:

\code{@describeIn foo see \\code{\\link{acutils::metamethod}}}
}

