% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun-multiplot.R
\name{multiplot}
\alias{multiplot}
\title{Multiple plot function}
\usage{
multiplot(..., plotlist = NULL, file, cols = 1, layout = NULL)
}
\arguments{
\item{...}{plot objects}

\item{plotlist}{list of plot objects}

\item{file}{path to file to print the plots into (?)}

\item{cols}{number of columns to arrange the plots into}

\item{layout}{matrix specifying the layout as explained in \code{\link{graphics::layout}};
if present, \code{cols} is ignored}
}
\description{
Print multiple ggplot plots on a single page. This is NOT splitting a single plot.
Unlike with facetting, the plots can be completely different.
}
\section{Layout}{

Layout specification by matrix works exactly like with graphics::layout.
If the layout is something like matrix(c(1, 2, 3, 3), nrow = 2, byrow = TRUE),
then plot 1 will go in the upper left, 2 will go in the upper right, and
3 will go all the way across the bottom.
}

\section{Reference}{

\href{http://www.cookbook-r.com/Graphs/
Multiple_graphs_on_one_page_(ggplot2)/}{R Graphics Cookbook} by Winston Chang
}

