% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ACutils_export.R
\name{mybarplot2}
\alias{mybarplot2}
\title{mybarplot2}
\usage{
mybarplot2(mydata, myexposure = NULL, myoutcome)
}
\arguments{
\item{mydata}{\link{tibble} the dataset}

\item{myexposure}{\link{string} the name of the categorical variable that defines the different levels.
If it is not NULL, different bar levels are produced.}

\item{myoutcome}{\link{string} the name of the categorical variable whose counts must the shown in a barplot.}
}
\value{
One bar for each level of level of the outcome variable. If there is exposure, an additional bar for each
level of the exposure is produced.
}
\description{
Useful function to produce barplot within for loops.
}
\examples{
mybarplot2(mydata = data2, myexposure =  "treatment", myoutcome =  "race")
}
