% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ACutils_export.R
\name{myboxplot2}
\alias{myboxplot2}
\title{myboxplot2}
\usage{
myboxplot2(mydata, myexposure = NULL, myoutcome)
}
\arguments{
\item{mydata}{\link{tibble} the dataset}

\item{myexposure}{\link{string} the name of the numerical variable whose boxplot must the produced.}

\item{myoutcome}{\link{string} the name of the categorical variable that defines the different levels. One boxplot for each level.}
}
\description{
Useful function to produce boxplot within for loops.
}
\examples{
myboxplot2(mydata = data, myexposure =  "treatment", myoutcome =  "age")
}
