% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-pipe.R
\name{\%>\%}
\alias{\%>\%}
\title{Pipe operator}
\usage{
lhs \%>\% rhs
}
\description{
Forward piping operator.
}
\details{
The value of the left hand side expression is passed to the right hand side expression.
It is matched to the first argument by default. This can be controlled by using the
dummy argument \code{.}, which stands in for the LHS value.

See package \code{\link[magrittr]{magrittr}} for details.
}
\keyword{internal}
