% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun-empty.plate.R, R/fun-plates.R,
%   R/fun-pmatrix.R
\name{empty.plate}
\alias{empty.plate}
\alias{plates}
\alias{pmatrix}
\title{plate functions}
\usage{
empty.plate(wells = 384)

pmatrix(x, rows, cols)
}
\arguments{
\item{wells}{number of wells in layout}

\item{x}{layout object: data frame, list or atomic vector}

\item{rows, cols}{backup for supplying non-standard matrix dimensions}
}
\value{
a data frame or a NULL if the format is not supported
}
\description{
Create and examine plate layouts.
}
\details{
\code{empty.plate} creates a scaffold: a data frame with well, row and column variables.
More variables can be added to is with a series of ifelse statements.
\code{pmatrix} can then conveniently display all features in matrix form.
}
\section{Functions}{
\itemize{
\item \code{empty.plate}: create a layout scaffold

\item \code{pmatrix}: quick peek at a layout matrix
}}

\section{Scaffold}{

The scaffold is a data frame with three columns: well, row, and column.
Rows and columns are numbered as the given format requires, e.g.
a 384-well plate has rows A through P and columns 1 through 24.
}

\section{Formats}{

This function is only so smart.
It has some hard coded formats that it will recognize, otherwise it returns a NULL
\itemize{
\item{* full plates of 6, 12, 24, 48, 96 and 384 wells,}
\item{* partial 384 plates:}
\itemize{
\item{336 wells: only rows B through O}
\item{228 wells: only rows C through N}
}
}
}

\section{Feature display}{

Each feature of the plate is represented as a matrix corresponding to the plate
dimensions. The format is determined based on the well number.

\code{pmatrix} accepts a single feature (vector) or a whole layout object
(a data frame) or a list.

The well number is rigidly determined based on the dimensions of x
then a wrapper for matrix() is created and applied to all columns
}

