% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ACutils_export.R
\name{plot_density}
\alias{plot_density}
\title{Plot the desity function}
\usage{
plot_density(
  ...,
  dist,
  main = " ",
  x_label = " ",
  y_label = " ",
  xlim = NULL,
  ylim = NULL,
  col = "gray48",
  use_x11_device = TRUE,
  use_ggplot = FALSE,
  add = FALSE
)
}
\arguments{
\item{...}{the list of parameters needed by the distribution specified in dist.}

\item{dist}{string with the name of the density to be plotted. Possibilities are,
\code{"norm"} (requires mean and sd), \code{"gamma"} (requires shape and rate), \code{"inv-gamma"} (requires shape and scale, see below), \code{"chi-sq"} (requires dof), \code{"exp"} (requires rate),
\code{"t"} (requires dop), \code{"Fisher"} (requires dof1 and dof2), \code{"nct"} (requires degree of freedom, location and scale. See \code{\link{dnct}} for the definition of such a distribution.).
\code{"laplace"} (requires location and scale).
The \code{"inv-gamma"} distribution is defined as follow:
\mjsdeqn{X~\sim~inv-gamma(shape = \alpha, scale = \beta)} if
\mjtdeqn{$$\begin{eqnarray*}f(X|\alpha,\beta) = \frac{\beta^{\alpha}}{\Gamma(\alpha)}\left(\frac{1}{x}\right)^{\alpha+1}e^{- \frac{\beta}{x}}{eqnarray*}$$}{\begin{eqnarray*}f(X|\alpha,\beta) = \frac{\beta^{\alpha}}{\Gamma(\alpha)}\left(\frac{1}{x}\right)^{\alpha+1}e^{- \frac{\beta}{x}}\end{eqnarray*}}{\begin{eqnarray*} f(X|\alpha,\beta) = \frac{\beta^{\alpha}}{\Gamma(\alpha)}\left(\frac{1}{x}\right)^{\alpha+1}e^{- \frac{\beta}{x}}\end{eqnarray*}}}

\item{main}{the title of the plot.}

\item{x_label}{the name of the x-axis in the plot.}

\item{y_label}{the name of the y-axis in the plot.}

\item{xlim}{set the range of x-axis. Not used if add is TRUE.}

\item{ylim}{set the range of y-axis. Not used if add is TRUE.}

\item{col}{set color of density line.}

\item{use_x11_device}{boolean, if x11() device has to be activated or not.}

\item{add}{boolean, if current plot has to be added to previous plot or not.}
}
\value{
this function does not return anything
}
\description{
\loadmathjax This function plots the density function of some know distributions. Current version, does not allow to overlap different plots
if use_ggplot is active. Though, it is possible when using standard plots. In such a case, the legend can be added a posteriori.
}
