% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun-print_plots.R
\name{print_plots}
\alias{print_plots}
\title{print some plots to a pdf file}
\usage{
print_plots(file = "some_plots.pdf", pattern = "plot\\\\.[0-9]+", ...)
}
\arguments{
\item{pattern}{regular expression to locate plot objects by}

\item{...}{arguments passed to other functions,
specifically \code{width} and \code{height} to \code{pdf}}

\item{filename}{file to print to}
}
\description{
Gather multipel existing ggplot objects and print them to one file
(one plot per page).
}
\details{
This is a simple utility function. It finds all objects in the global
environment whose names match a given regular expression and tries to send
them to the pdf device. In an analysis where multiple plots are produced,
this can help to export them into neat files, with some care for naming them.

NULL objects will be ignored
}
