% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun-rename_reference_observations.R
\name{rename_reference_observations}
\alias{rename_reference_observations}
\title{rename reference observations}
\usage{
rename_reference_observations(
  X,
  In_Variable = "treatment",
  Replacee = "nt",
  Replacers,
  return.composite = TRUE
)
}
\arguments{
\item{X}{data frame}

\item{In_Variable}{name of the variable whose values are to be altered (character)}

\item{Replacee}{name of the value of \code{In_Variable} that is to be replaced (character)}

\item{Replacers}{character vector of values that will replace \code{Replacee};
defaults to all other values of \code{In_Variable}}

\item{return.composite}{logical flag whether to return just only reference rows
or the whole data set (with references renamed)}
}
\value{
A data frame. If \code{return.composite} is \code{FALSE} only the (renamed)
observations will be returned. It \code{TRUE}, the returned data frame
contains the renamed reference observations as well as the original non-reference observations.
}
\description{
Change category descriptions in a selected variable to enable prettier plotting.
}
\details{
Experimental data are kept in a data frame and some observations
constitute a reference group. In order to be able to show them alongside
test observations, they may need to have their categorical variables altered
e.g. non-treated cells can be renamed to cells treated for 0 hours.

The categorical variable can be stored as character or as a factor
and this will be preserved but the replaced level will be dropped.

Grouping (by \code{dplyr}) will be preserved.
}
