% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun-rescale.R
\name{rescale}
\alias{rescale}
\alias{rescale.default}
\alias{rescale.matrix}
\title{scale numeric vector}
\usage{
rescale(x, range)

\method{rescale}{default}(x, range)

\method{rescale}{matrix}(x, range)
}
\arguments{
\item{x}{a numeric vector or matrix}

\item{range}{range which to scale to:
\code{positive}: [0,1], \code{negative}: [-1,0], \code{both}: [-1,1]}
}
\value{
a numeric object of the same class as \code{x}
}
\description{
Scales a vector to a requested range.
}
\section{Methods (by class)}{
\itemize{
\item \code{default}: defult method for vectors

\item \code{matrix}: calls default method and copies the
\code{dim} attribute of \code{x} to result
}}

