% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ACutils_export.R
\name{rnct}
\alias{rnct}
\title{Non Central Student t - Random Number Generator}
\usage{
rnct(n = 1, n0, mu0, gamma0)
}
\arguments{
\item{n}{the number of points to be generated.}
}
\value{
scalar generated from Non Central Student t.
}
\description{
This function generates a random number distributed as Non Central Student t. See \code{\link{dnct}} for the definition of such a distribution.
}
