% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun-show_colors.R
\name{show_colors}
\alias{show_colors}
\title{display shades of a color}
\usage{
show_colors(color)
}
\arguments{
\item{color}{a regular expression for color determination}
}
\description{
Quick peek at available variations of a color in R.
}
\details{
The function uses the \code{color} argument as a regular expression
with which to search \code{colors()} and displays all matching colors
on a rudimentary barplot.
}
