% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun-sphere.R
\name{sphere}
\alias{sphere}
\title{compute sphere volume}
\usage{
sphere(r, n = 7L)
}
\arguments{
\item{r}{radius}

\item{n}{number of dimensions}
}
\value{
volume of a supersphere
}
\description{
Calculate volume of n-dimentional sphere of radius r.
}
