% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun-split_input_string.R
\name{split_input_string}
\alias{split_input_string}
\title{extract substrings from text input}
\usage{
split_input_string(x, sep = ";:,&+=/ ")
}
\arguments{
\item{x}{a character string}

\item{sep}{character string containing allowed separators}
}
\value{
A character vector of isolated substrings.
}
\description{
Isolates elements from a string.
}
