% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ACutils_export.R
\name{t_col}
\alias{t_col}
\title{Makes colors trasparent}
\usage{
t_col(color, percent = 30, name = NULL)
}
\arguments{
\item{color}{the color name}

\item{percent}{percentage of transparecy. Low values stands for more transparecy.}

\item{name}{an optional name for the new color}
}
\value{
the new color, which is also saved.
}
\description{
Get a color and returns its transparent version. New color can be named.
}
