% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ACutils_export.R
\name{upper2complete}
\alias{upper2complete}
\title{Upper to complete matrix}
\usage{
upper2complete(A)
}
\arguments{
\item{A}{an upper triangular matrix. Lower part is not used and overwritten}
}
\value{
symmetric matrix
}
\description{
Maps upper triangular matrix into symmetric matrix. Keeps the same diagonal.
}
