% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ACutils_export.R
\name{boxplot_from_vectors}
\alias{boxplot_from_vectors}
\title{Plot boxplot from vectors}
\usage{
boxplot_from_vectors(
  v1,
  v2,
  v3,
  use_x11_device = TRUE,
  use_ggplot = TRUE,
  title = "Title",
  x_label = "x_axis",
  y_label = "y_axis"
)
}
\value{
this function does not return anything
}
\description{
Plot boxplot from vectors
}
