% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun-insert_zeros.R
\name{insert_zeros}
\alias{insert_zeros}
\alias{insert_zeros.character}
\alias{insert_zeros.factor}
\alias{insert_zeros.numeric}
\alias{insert_zeros.default}
\title{insert zeros into a character vector}
\usage{
insert_zeros(x, zeros = "auto", after = 1)

\method{insert_zeros}{character}(x, zeros = "auto", after = 1)

\method{insert_zeros}{factor}(x, zeros = "auto", after = 1)

\method{insert_zeros}{numeric}(x, zeros = "auto", after = 1)

\method{insert_zeros}{default}(x)
}
\arguments{
\item{x}{character vector}

\item{zeros}{number of \code{0} characters to insert; see \code{Details}}

\item{after}{position of the character after which the zeros will be inserted}
}
\value{
Modified character vector, where strings of zeros have been inserted at the specified position.
}
\description{
Insert a string of zeros into every item of a character vector.
\code{A1} becomes \code{A01}.
The orginal intent was to fix sorting of strings where \code{2 > 12} but \code{02 < 12}.
More robust options were built in just in case.
}
\details{
The \code{zeros} argument can be a fixed number but it defaults to \code{auto}. Then:
if \code{X} is of length 1, adds 1 \code{0};
if \code{X} is longer than 1, adds to each element a number of \code{0}
so that all elements of \code{X} have an equal number of characters.
}
\section{Methods (by class)}{
\itemize{
\item \code{character}: the basic method

\item \code{factor}: runs on levels of x rather than its body

\item \code{numeric}: coerces to character and passes to character method

\item \code{default}: throws error for non-supperted classes
}}

