% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun-rescale.R
\name{rescale}
\alias{rescale}
\alias{rescale.default}
\alias{rescale.matrix}
\title{scale numeric vector}
\usage{
rescale(x)

\method{rescale}{default}(x)

\method{rescale}{matrix}(x)
}
\arguments{
\item{x}{a numeric vector or matrix}
}
\value{
a numeric vector object of the same class as \code{x}
}
\description{
Scales a vector. Positive values are scaled from 0 to 1, negative ones are scaled -1 to 0.
}
\section{Methods (by class)}{
\itemize{
\item \code{default}: defult method for vectors

\item \code{matrix}: calls default method and copies the
\code{dim} attribute of \code{x} to result
}}

