% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun-rescale.R
\name{rescale}
\alias{rescale}
\alias{rescale.default}
\alias{rescale.matrix}
\title{scale numeric vector}
\usage{
rescale(x, range = c("positive", "negative", "both"))

\method{rescale}{default}(x, range = c("positive", "negative", "both"))

\method{rescale}{matrix}(x, range = c("positive", "negative", "both"))
}
\arguments{
\item{x}{a numeric vector or matrix}

\item{range}{range which to scale to}
}
\value{
a numeric object of the same class as \code{x}
}
\description{
Scales a vector to a requested range.
}
\details{
Values in the vector are scaled such to fill \code{range}:
\itemize{
  \item{\code{positive}: from 0 to 1
  }
  \item{\code{negative}: from -1 to 0
  }
  \item{\code{both}: from -1 to 1
  }
}
}
\section{Methods (by class)}{
\itemize{
\item \code{default}: defult method for vectors

\item \code{matrix}: calls default method and copies the
\code{dim} attribute of \code{x} to result
}}

